Modernize env; replace bundled GPLv3 ANARCI with conda BSD ANARCI
The prediction path already ran on modern libraries; only two behavior-preserving
fixes were needed, and the vendored ANARCI is now an external BSD dependency.

Modernization (default env: python 3.11 / numpy 2.4 / pandas 3.0 / biopython 1.87
/ hmmer 3.4):
- utils.py: np.float(...) -> float literal (np.float removed in numpy >= 1.24;
  this is in the builddb path, not prediction).
- predict.py, prosci/util/pdb.py: raw-string regex literals (clears
  invalid-escape DeprecationWarnings; future-proofs py3.12).

ANARCI:
- Remove the vendored GPLv3 ANARCI (modules + HMM data) and depend on bioconda
  anarci 2024.05.21 (BSD-3, noarch). Verified it reproduces all 12 webserver
  golden combos exactly (Tier A sequences + Tier B canonical/median), so the
  numbering is unchanged.
- Imports now use `from anarci import run_anarci`; region_definitions.py
  (self-contained) is unchanged.
- setup.py / MANIFEST.in drop scalop.anarci; also fix MANIFEST's data->database
  path and the nonexistent BioPython-licence entry.

Environment:
- Retire the period-appropriate `legacy` env now that the baseline is frozen and
  reproduced by the modern env; pixi.toml is a single default env.

README: add anarci to the conda install line and dependency list.

pixi run test -> 26 passed.

+++ lib/python/scalop/utils.py
@@ -1,5 +1,5 @@
 from .region_definitions import Accept
-from .anarci import run_anarci
+from anarci import run_anarci
 
 import os, sys, time, re
 import pickle
@@ -53,7 +53,7 @@ def get_PSSM(storepos):
 	for pos in sorted(storepos):
 		for resi in range(20):
 			if storepos[pos][resi] == 0:
-				storepos[pos][resi] = np.float(0.001)
+				storepos[pos][resi] = 0.001
 			else:
 				storepos[pos][resi] /=float(nseq[pos])
 	pssm = {pos: np.log2(np.array(storepos[pos])/float(0.05)) for pos in sorted(storepos)}
